tourism_companion/requirements.txt
fastapi
uvicorn
pycld2
langid
openai
transformers
torch
sqlalchemy
pydantic
whisper
pdfplumber
faiss-cpu
langchain
